Code review fixes
- Add further comments
- Renames to make things clearer
- Move to primary constructor
- Spelling fixes
- Rename file to match class name
